Make dock role idempotent
This only works if the dock entry list contains the same URI for each
item that would be output by `dockutil --list`. For example, if the list
contains:

    - name: Firefox
      uri: /Applications/Firefox.app

Then the dock will be changed, even if it contains the following entry:

    - name: Firefox
      uri: file:///Applications/Firefox.app/

The two must match exactly for the update to be skipped.